fix(apache-composer): help editor find openwhisk-composer module by moving appModulePath to apache-composer preload
Fixes #317
