+++ plugins/plugin-apache-composer/src/test/composer1/composer-viz.ts
@@ -51,6 +51,7 @@ const requireAbsolute = composerInput('require-absolute.js')
 const requireRelative = composerInput('require-relative.js')
 const fsRead = composerInput('fs-read.js')
 const addSubscription = composerErrorInput('addSubscription.js')
+const owComposerErr = composerErrorInput('openwhisk-composer-throw-err.js')
 
 /**
  * Here starts the test
@@ -61,6 +62,10 @@ describe('show the composer visualization without creating openwhisk assets', fu
   after(common.after(this))
 
   it('should have an active repl', () => cli.waitForRepl(this.app))
+
+  it('should show error thrown by openwhisk-composer node_module', () => cli.do(`preview ${owComposerErr.path}`, this.app)
+    .then(cli.expectError(0, `no such file or directory, open 'doesnotexist.js'`))
+    .catch(common.oops(this)))
 
   it('should preview an empty composition', () => cli.do(`app preview data/composer/composer-source/empty.js`, this.app)
     .then(verifyTheBasicStuff('empty.js', 'composerLib'))
