+++ plugins/plugin-apache-composer/tests/data/composer/composer-source-expect-errors/openwhisk-composer-throw-err.js
@@ -1,0 +1,3 @@
+const composer = require('openwhisk-composer')
+
+module.exports = composer.sequence(composer.action('A', { filename: 'doesnotexist.js' }))
